bug fixes; ui improvements

+++ Rooster-Shared/Preferences/DataStore/PreferencesStorage.swift
@@ -9,58 +9,55 @@ import Foundation
 import RoosterCore
 import Cocoa
 
-public class PreferencesStorage {
+public class PreferencesStorage: Loggable {
     
     public enum Errors: Error {
         case failedToLoadApplicationSupportDirectory
     }
-    
-    private var cachedPreferences: Preferences?
     
     public let directoryURL: URL
     public let soundsSetsFolderURL: URL
     
     private let prefsFile: JsonFile<Preferences>
     
-    init() throws {
-        self.cachedPreferences = nil
+    private var createdFolders: Bool = false
     
-        guard let applicationSupportDirectory = FileManager.default.urls(for: .applicationSupportDirectory, in: .userDomainMask).first,
-           let bundleIdentifier = Bundle.main.bundleIdentifier else {
-            
-            throw Errors.failedToLoadApplicationSupportDirectory
-        }
+    init() {
+    
+        let applicationSupportDirectory = FileManager.default.applicationSupportDirectoryURL
         
-        let directoryURL = applicationSupportDirectory.appendingPathComponent(bundleIdentifier)
+        let directoryURL = applicationSupportDirectory.appendingPathComponent(Bundle.main.bundleIdentifier ?? "")
         let preferencesFileURL = directoryURL.appendingPathComponent("Preferences.json")
         let soundsSetsFolderURL = directoryURL.appendingPathComponent("Sound Sets")
-
-        try FileManager.default.createFolderIfNeeded(atURL: directoryURL)
-        try FileManager.default.createFolderIfNeeded(atURL: soundsSetsFolderURL)
 
         self.directoryURL = directoryURL
         self.soundsSetsFolderURL = soundsSetsFolderURL
         self.prefsFile = JsonFile<Preferences>(withURL: preferencesFileURL)
     }
     
-    func read() throws -> Preferences {
-        
-        if let cached = self.cachedPreferences {
-            return cached
+    private func createFoldersIfNeeded() throws {
+        if !self.createdFolders {
+            try FileManager.default.createFolderIfNeeded(atURL: self.directoryURL)
+            try FileManager.default.createFolderIfNeeded(atURL: self.soundsSetsFolderURL)
         }
-        
-        let prefs = try self.prefsFile.read()
-        self.cachedPreferences = prefs
-        return prefs
+    }
+    
+    func read() throws -> Preferences {
+        try self.createFoldersIfNeeded()
+        return try self.prefsFile.read()
     }
     
     func write(_ preferences: Preferences) throws {
+        try self.createFoldersIfNeeded()
         try self.prefsFile.write(preferences)
-        self.cachedPreferences = preferences
     }
 
     func delete() throws {
-        try self.prefsFile.delete()
+        if self.prefsFile.exists {
+            try self.prefsFile.delete()
+            self.logger.log("deleted pref file: \(self.prefsFile.url.path)")
+        }
+    
     }
     
     public var exists: Bool {
